Move database related functions to its own module
